seperated components into their own files
